requirements.txt
asgiref==3.8.1
Django==5.1.6
django-recaptcha==4.1.0
gunicorn==23.0.0
packaging==25.0
pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.3
